+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.6] - 2026-03-17
+- Fix color accessibility.
+## [1.0.5] - 2026-03-17
+- Apply clean code fix minors bugs.
 ## [1.0.4] - 2026-03-15
 - Apply Clean Architecture in Save class.
 - Refactor Sheet class.